忽略非操作的场景

+++ src/js/content.js
@@ -1,5 +1,3 @@
-console.log("content.js");
-
 window.addEventListener('load', function () {
     chrome.runtime.sendMessage({ type: "xsrf-token" }, function (response) {
         let token = response.token;
@@ -8,6 +6,10 @@ window.addEventListener('load', function () {
         }
 
         chrome.storage.local.get(['shortenUrl', 'shortenTabId'], function (result) {
+            if (!result.shortenUrl) {
+                return;
+            }
+
             chrome.storage.local.clear();
             shortenOnTab(result.shortenUrl, result.shortenTabId, token);
         });
